fix: Fix minor bugs / inconsistencies related to casting types, timezones, and data precision, add tests for `analytics.py`

+++ src/order_book_simulator/market_data/analytics.py
@@ -119,8 +119,8 @@ class MarketDataAnalytics:
             df.filter(pl.col("last_trade_price") != "")
             .with_columns(
                 [
-                    pl.col("last_trade_price").cast(pl.Decimal).alias("price"),
-                    pl.col("last_trade_quantity").cast(pl.Decimal).alias("quantity"),
+                    pl.col("last_trade_price").cast(pl.Float64).alias("price"),
+                    pl.col("last_trade_quantity").cast(pl.Float64).alias("quantity"),
                 ]
             )
             .with_columns((pl.col("price") * pl.col("quantity")).alias("price_volume"))
